feat: implimenting models functionally
