+++ main.go
@@ -5,25 +5,10 @@ import (
 	"log"
 	"net/http"
 
+	db "github.com/wh-iterabb-it/catfacts-api/models"
+
 	"github.com/gorilla/mux"
 )
-
-// a fun* fact about cats
-// * = many facts are not fun
-type CatFact struct {
-	ID     string  `json:"id,omitempty"`
-	Fact   string  `json:"fact,omitempty"`
-	Source *Source `json:"source,omitempty"`
-}
-
-// where the hell did this fact come from?
-type Source struct {
-	Name string `json:"name,omitempty"`
-	Url  string `json:"url,omitempty"`
-}
-
-// catfacts is a collection of CatFact
-var catfacts []CatFact
 
 // GetCatFactsEndpoint is used for getting a collection of catfacts
 func GetCatFactEndpoint(w http.ResponseWriter, req *http.Request) {
@@ -43,8 +28,7 @@ func GetCatFactEndpoint(w http.ResponseWriter, req *http.Request) {
 
 // GetCatFactsEndpoint is used for getting a collection of catfacts
 func GetCatFactsEndpoint(w http.ResponseWriter, req *http.Request) {
-	// encodes and returns array of catfacts
-	json.NewEncoder(w).Encode(catfacts)
+	db.getCatFacts(w)
 }
 
 // CreateCatFactEndpoint is used for creating a new catfact
@@ -71,6 +55,8 @@ func DeleteCatFactEndpoint(w http.ResponseWriter, req *http.Request) {
 	}
   json.NewEncode(w).Encode(catfacts)
 }
+
+func AddCatFact()
 
 func main() {
 	router := mux.NewRouter()
